requirements.txt
Django==4.2.10
django_environ==0.4.5
openai==1.35.3
python-dotenv==1.0.1
